getCartItems done

+++ src/CartItem.java
@@ -1,8 +1,8 @@
 public class CartItem {
   public Item item;
-  public float count;
+  public double count;
 
-  public CartItem(Item item, float count) {
+  public CartItem(Item item, double count) {
     this.item = item;
     this.count = count;
   }
